refactor(haskell): break down into modules
