+++ haskell/related-post-gen.cabal
@@ -5,24 +5,58 @@ author:        thelissimus
 maintainer:    thelissimus@tuta.io
 build-type:    Simple
 
-executable related-post-gen
+common warnings
   ghc-options:
     -Wall -Wextra -Weverything -Wcompat -Wno-implicit-prelude
     -Wno-unsafe -Wno-missing-safe-haskell-mode
     -Wno-missing-local-signatures -Wno-missing-import-lists
-    -Wno-all-missed-specialisations -O2
+    -Wno-all-missed-specialisations
 
-  main-is:          Main.hs
+library
+  import:           warnings
+  ghc-options:      -O2
+  hs-source-dirs:   src
+  exposed-modules:  RelatedPostGen
   build-depends:
     , aeson
-    , base               ^>=4.17.2.0
-    , bytestring
+    , base               >=4.7 && <5
     , deepseq
     , text
-    , time
     , unboxed-ref
     , vector
     , vector-hashtables
 
+  default-language: GHC2021
+
+executable related-post-gen
+  import:           warnings
+  ghc-options:      -O2
   hs-source-dirs:   app
+  main-is:          Main.hs
+  build-depends:
+    , aeson
+    , base              >=4.7 && <5
+    , bytestring
+    , deepseq
+    , related-post-gen
+    , time
+    , vector
+
+  default-language: GHC2021
+
+benchmark related-post-gen-bench
+  import:           warnings
+  ghc-options:      -with-rtsopts=-A32m -fproc-alignment=64
+  type:             exitcode-stdio-1.0
+  main-is:          Main.hs
+  hs-source-dirs:   bench
+  build-depends:
+    , aeson
+    , base              >=4.7 && <5
+    , bytestring
+    , deepseq
+    , related-post-gen
+    , tasty-bench
+    , tasty-papi
+
   default-language: GHC2021
